Feature/api v146 and missing membes up to 2024-05-15

+++ Meraki.Api/Data/LiveToolsThroughputTestCreateResponseResultSpeeds.cs
@@ -12,5 +12,5 @@ public class LiveToolsThroughputTestCreateResponseResultSpeeds
 	/// </summary>
 	[ApiAccess(ApiAccess.Read)]
 	[DataMember(Name = "downstream")]
-	public int Downstream { get; set; }
+	public double Downstream { get; set; }
 }
